third_party/blink/web_tests/external/wpt/html/rendering/non-replaced-elements/form-controls/resets-expected.txt

This is a testharness.js-based test.
Found 1 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] <textarea> - overflow
  assert_equals: expected "visible" but got "auto"
Harness: the test ran to completion.
